+++ server/catgenome/src/main/java/com/epam/catgenome/manager/blast/dto/RequestInfoPayload.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class RequestInfoPayload {
-    private int requestId;
+    private long requestId;
     private String status;
     private String taskType;
     private Date createdDate;
